Add files via upload
The purpose of this project is to create an employee/project management system for enterprise use that Software Development companies of up to medium size can deploy in order to keep track of employee information (such as date joined, personal information, salary, etc.), what project employees are currently working on, and performance reviews by superiors. 

This project will be created using C++ and Pqxx to connect to a PostgreSQL database. It will be delivered in the form of an interactive terminal application. 
We  separate the whole system to 4 main tables. Manager and Developers tables inherit Employee with all personal information.
Project contains their information and has relationships with the Manager (in Supervise) and Developers (through WorkOn). Finally, the Performance table allows manager grades for their developers.